ai-upscaler/celery_app.py
from celery import Celery
from config import Config
import boto3

celery_app = Celery(
    'ai_upscaler',
    broker=Config.CELERY_BROKER_URL,
    backend=Config.CELERY_RESULT_BACKEND
)

@celery_app.task
def upscale_image_task(job_id: str, s3_input_key: str):
    """
    Celery task to upscale an image
    """
    try:
        # Initialize S3 client
        s3_client = boto3.client(
            's3',
            endpoint_url=Config.AWS_ENDPOINT_URL,
            aws_access_key_id=Config.AWS_ACCESS_KEY_ID,
            aws_secret_access_key=Config.AWS_SECRET_ACCESS_KEY,
            region_name=Config.AWS_DEFAULT_REGION
        )
        
        # Download image from S3
        # Process with AI model (placeholder)
        # Upload result to output bucket
        
        output_key = f"output/{job_id}/upscaled.jpg"
        
        return {
            "job_id": job_id,
            "status": "completed",
            "output_key": output_key
        }
        
    except Exception as e:
        return {
            "job_id": job_id,
            "status": "failed",
            "error": str(e)
        }