ai-upscaler/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
boto3==1.34.0
pika==1.3.2
redis==5.0.1
pillow==10.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
requests==2.31.0
prometheus-client==0.19.0


